requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-cacheops==4.1
django-debug-toolbar==1.10.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
funcy==1.11
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.3.0
psycopg2==2.7.5
Pygments==2.2.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
